backend/requirements.txt
# ── Django & REST ──────────────────────────────────────────────────────────────
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
django-filter==24.2
django-cors-headers==4.3.1

# ── Database ───────────────────────────────────────────────────────────────────
psycopg2-binary==2.9.9

# ── ASGI Server ──────────────────────────────────────────────────────────────
uvicorn[standard]==0.30.1
gunicorn==22.0.0

# ── Cache / Channels / WebSockets ─────────────────────────────────────────────
channels==4.1.0
channels-redis==4.2.0
django-redis==5.4.0
redis==5.0.6

# ── Celery (background tasks) ──────────────────────────────────────────────────
celery==5.4.0
django-celery-beat==2.6.0

# ── Static files ───────────────────────────────────────────────────────────────
whitenoise==6.7.0
Pillow==10.4.0

# ── Config ─────────────────────────────────────────────────────────────────────
python-decouple==3.8

# ── HTTP (for Termii / SendGrid / AXpress proxy / Coach AI) ─────────────────
requests==2.32.3
httpx==0.28.1

# ── Firebase (rider push) ──────────────────────────────────────────────────────
firebase-admin==6.5.0
